+++ src/lib.rs
@@ -24,6 +24,7 @@ use dioxus_document as document;
 use browserinfo::{broinfo_js, BroInfo, Browser};
 
 # async fn func() -> anyhow::Result<()> {
+use browserinfo::FromJsonStr;
 let js_bro: &str = broinfo_js();
 let eval = document::eval(js_bro).await?;
 let json_str = eval.to_string();